# GitRepoMind/server/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "GitRepoMind API",
                 "description":  "AI-powered GitHub Repository Analyzer \u0026 RAG Chat",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/repos/{repo_id}/stats":  {
                                                     "get":  {
                                                                 "tags":  [
                                                                              "repos"
                                                                          ],
                                                                 "summary":  "Get Repository Stats",
                                                                 "description":  "Get statistics for an indexed repository.\n\nCombines cached analysis data with vector store statistics\nto provide comprehensive repository information.\n\nArgs:\n    repo_id: Repository ID returned from analyze-repo endpoint\n\nReturns:\n    RepoStatsResponse with detailed repository statistics\n\nRaises:\n    HTTPException: If repository not found or operation fails",
                                                                 "operationId":  "get_repository_stats_api_repos__repo_id__stats_get",
                                                                 "parameters":  [
                                                                                    {
                                                                                        "name":  "repo_id",
                                                                                        "in":  "path",
                                                                                        "required":  true,
                                                                                        "schema":  "@{type=string; title=Repo Id}"
                                                                                    }
                                                                                ],
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "Successful Response",
                                                                                               "content":  "@{application/json=}"
                                                                                           },
                                                                                   "422":  {
                                                                                               "description":  "Validation Error",
                                                                                               "content":  "@{application/json=}"
                                                                                           }
                                                                               }
                                                             }
                                                 },
                  "/api/repos/{repo_id}":  {
                                               "delete":  {
                                                              "tags":  [
                                                                           "repos"
                                                                       ],
                                                              "summary":  "Delete Repository",
                                                              "description":  "Delete an indexed repository from cache and vector store.\n\nRemoves all associated data:\n- Vectors from Qdrant\n- Cached analysis results\n- Conversation history (if exists)\n\nArgs:\n    repo_id: Repository ID returned from analyze-repo endpoint\n\nReturns:\n    Empty response on success\n\nRaises:\n    HTTPException: If repository not found or operation fails",
                                                              "operationId":  "delete_repository_api_repos__repo_id__delete",
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "repo_id",
                                                                                     "in":  "path",
                                                                                     "required":  true,
                                                                                     "schema":  "@{type=string; title=Repo Id}"
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "204":  {
                                                                                            "description":  "Successful Response"
                                                                                        },
                                                                                "422":  {
                                                                                            "description":  "Validation Error",
                                                                                            "content":  "@{application/json=}"
                                                                                        }
                                                                            }
                                                          }
                                           },
...
              },
...
}